Fixed some warnings.

+++ src/post_modules/Nettle/idea.c
@@ -83,8 +83,8 @@ inv(unsigned INT16 x)
 
   if (x <= 1)
     return x;	/* 0 and 1 are self-inverse */
-  t1 = 0x10001L / x;	/* Since x >= 2, this fits into 16 bits */
-  y = 0x10001L % x;
+  t1 = (unsigned INT16) (0x10001L / x); /* Since x >= 2, this fits into 16 bits */
+  y = (unsigned INT16) (0x10001L % x);
   if (y == 1)
     return low16(1-t1);
   t0 = 1;
